doc/bug: Fixing 404 errors in docs
Recent changes to the docs were using incorrect filenames in the
documentation (missing the `.md` file extention which is required for
files beyond `README.md` as the default in a directory).

doc: Updating main README.md, removing stale data

Previously the main README.md in the project referenced a sample python
server which is no longer a part of the project.  Additionally, there is
now a more clearly codified location for documentation in the project.