device: Wiegand badge reader support
This commit provides support for [Wiegand][1] style badge readers.  It has
been tested with a number of EM4100 style readers and has been used in
production for a number of months without issue.

The wiring for a badge reader requires two data GPIO pins as input.
These pins within the Wiegand protocol are "D0" and "D1" respectively.
Normally, these pins are pulled high (generally to 5v).  They perform
differential signalling through pulling D0 low to represent a "0" and
pulling D1 low to represent a "1".

This introduces some challenges as most of the connectors have only a
single raw GPIO input.  In practice, this has been used with J5
consuming the B5_LOGIC and B6_LOGIC pins (pin 13 and 7 respectively on
version 1.0 of the Makerspace Auth board).

Within the code this bitstream of "1"s and "0"s is represented as a
string so as to avoid additional dependencies as well as to not worry
about specific card formats.  For our purposes, this module should
pedantically output the hexadecimal value and leave card ids, facility
codes, and other format specific nuances to other components of
software.

[1]: https://en.wikipedia.org/wiki/Wiegand_interface